mention OS
I tried to start modding on my Linux machine, which was naive, I know. But the mention of a "Terminal" which was not specified as "PowerShell" made me optimistically curious. When I realized the Window for the BepInEx console can't even be properly spawned when the game is run with Proton, I stopped and thought I suggest this little addition.